+++ azext_iot/digitaltwins/common.py
@@ -23,6 +23,7 @@ LRO_TIMER = 60
 
 
 # Data History strings
+ADX_DEFAULT_TABLE = "AdtPropertyEvents"
 DT_INSTANCE = "Digital Twins instance"
 USER_ASSIGNED_IDENTITY = "User Assigned Identity"
 DT_SYS_IDENTITY_ERROR = "Digital Twins instance does not have System-Assigned Identity enabled. Please enable and try again."